refactor: Move simulation to extra module; move exports to prelude
