+++ src/resource.rs
@@ -10,8 +10,7 @@ use itertools::Itertools;
 use backtrace::Backtrace as trc;
 use thiserror::Error;
 
-use crate::units::*;
-use crate::*;
+use crate::prelude::*;
 
 #[derive(Clone, PartialEq, Eq, Hash, Debug, Display, Default, From, AsRef, AsMut, Into)]
 pub struct ResourceName(String);
